+++ Production/govuk_ios/Extensions/UIKit/UITabBarAppearance+GOVUK.swift
@@ -5,7 +5,7 @@ extension UITabBarAppearance {
     static var govUK: UITabBarAppearance {
         let appearance = UITabBarAppearance()
         appearance.configureWithTransparentBackground()
-        appearance.shadowColor = UIColor.primaryDivider
+        appearance.shadowColor = UIColor.govUK.strokes.listDivider
         appearance.backgroundColor = UIColor.govUK.fills.surfaceFixedContainer
         appearance.inlineLayoutAppearance = .govUK
         appearance.stackedLayoutAppearance = .govUK
